rm example fsdp
