+++ chatlearn/algorithm/grpo.py
@@ -14,9 +14,14 @@
 # ==============================================================================
 """grpo algorithm"""
 
+import json
+import random
 from dataclasses import dataclass, field
-from typing import Any, Optional, List
-# from chatlearn.configs.common import (
+from typing import Any, Optional, List, Dict
+from collections import defaultdict
+
+import torch
+
 from configs.common import (
     BaseConfig,
     RuntimeEnvConfig,
@@ -26,9 +31,19 @@ from configs.common import (
     PolicyTrainerConfig,
     BaseModelConfig
 )
-# from chatlearn.algorithm.base_algo import BaseAlgorithm
 from algorithm.base_algo import BaseAlgorithm
-from examples.fsdp.entry.train_grpo import *
+
+from chatlearn.algorithm.grpo_utils.policy_trainer import PolicyTrainer
+from chatlearn.algorithm.grpo_utils.vllm_policy_inference import VLLMPolicyInference
+from chatlearn.models.reward.rule_reward import RuleReward
+from chatlearn.runtime.environment import Environment
+from chatlearn.runtime.trainer import Trainer
+from chatlearn.runtime.evaluator import Evaluator
+from chatlearn.utils.utils import listdict_to_dictlist
+import chatlearn
+from chatlearn import Engine
+from chatlearn.models.base_module import BaseModule
+# from examples.fsdp.entry.train_grpo import *
 
 @dataclass
 class GrpoModelConfig(BaseConfig):
@@ -65,7 +80,82 @@ class GrpoConfig(BaseConfig):
         metadata={"help": "Grpo model config."}
     )
 
+def read_data_path_list(data_path_list: List[str], mode: str = "jsonl"):
 
+    data = []
+    for data_path in data_path_list:
+        if mode == "json":
+            with open(data_path, 'r') as f:
+                data.extend(json.load(f))
+        elif mode == "jsonl":
+            with open(data_path, 'r') as f:
+                data.extend([json.loads(line) for line in f])
+    return data
+
+def compute_grpo_adv(episode_replay_buffers):
+    buffers = episode_replay_buffers[-1].buffer
+    queryids2samples = defaultdict(list)
+    for s in buffers:
+        queryids2samples[hash(','.join(map(str, s["prompt_token_ids"])))].append(s)
+    
+    res_buffers = []
+    for _, l in queryids2samples.items():
+        rewards = [each["rule_rewards"] for each in l]
+        rewards = torch.cat(rewards, dim=0)
+
+        mean = torch.mean(rewards)
+        std = torch.std(rewards)
+        for i, li in enumerate(l):
+            li['advantages'] = ((rewards[i] - mean) / (std + 1e-5))
+        res_buffers.extend(l)
+    return res_buffers
+
+class GRPOEvaluator(Evaluator):
+
+    def post_process(self, results, eval_info):
+        # results Dict[List]
+        results = results["reward"]
+        results = listdict_to_dictlist(results)
+        # convert list[tensor(n,1)] to list[float]
+        rule_rewards = results.get("rule_rewards", [])
+        rule_rewards_flatten = torch.cat(rule_rewards).squeeze().tolist()
+
+        reward_score = sum(rule_rewards_flatten) / len(rule_rewards_flatten)
+        eval_reward_stats = {"eval_reward_score": reward_score}
+
+
+        self._metric_list.append(eval_reward_stats)
+        
+        return results
+
+
+class GRPOEngine(Engine):
+    """GRPO Engine."""
+    def __init__(self,
+                 policy: VLLMPolicyInference,
+                 reward: RuleReward,
+                 ref_policy: PolicyTrainer,
+                 policy_trainer: PolicyTrainer):
+        def env_compute_flow(batch):
+            policy_out = policy.forward_step(batch)
+            old_logprobs_out = policy_trainer.forward_step(policy_out)
+            ref_logprobs_out = ref_policy.forward_step(old_logprobs_out)
+            reward_out = reward.forward_step(ref_logprobs_out)
+            return ref_logprobs_out, reward_out
+
+        def trainer_compute_flow(batch):
+            policy_trainer.train_step(batch)
+
+        def evaluator_flow(batch):
+            policy_out = policy.forward_step(batch)
+            reward_out = reward.eval_forward(policy_out)
+            return reward_out
+
+        env = Environment(env_compute_flow)
+        trainer = Trainer(trainer_compute_flow)
+        evaluator = GRPOEvaluator(evaluator_flow)
+        super().__init__(environment=env, trainer=trainer, evaluator=evaluator, name='grpo')
+        self.set_parameter_sync(policy_trainer, policy)
 
 class GrpoAlgorithm(BaseAlgorithm):
     """GrpoAlgorithm"""
