--- examples/fsdp/configs/grpo/base.yaml
@@ -1,1 +1,0 @@
-seed: 1234
