--- examples/fsdp/configs/grpo/policy_trainer.yaml
@@ -1,10 +1,0 @@
-includes:
-  - base.yaml
-pretrain_or_model: ${model_path}
-learning_rate: 2e-6
-grad_clip: 1
-gradient_checkpointing: True
-# for grpo algorithm
-pos_clip_ratio: 0.2
-negative_clip_ratio: 0.2
-save_hf: True
