--- examples/fsdp/configs/grpo/reference.yaml
@@ -1,3 +1,0 @@
-includes:
-  - base.yaml
-pretrain_or_model: ${model_path}
