+++ grpo.yaml
@@ -3,15 +3,15 @@ runtime_args:
   exp_name: grpo_test
   colocation: [policy,policy_trainer,ref_policy]
   # path config
-  output_dir: /mnt/workspace/yanhaiqiang.yhq/parameter_refactor/ChatLearn/output/grpo_test
-  data_path: /mnt/workspace/yanhaiqiang.yhq/parameter_refactor/ChatLearn/dataset/MATH-lighteval/train.json
-  eval_data_path: /mnt/workspace/yanhaiqiang.yhq/parameter_refactor/ChatLearn/dataset/MATH-lighteval/test.json
+  output_dir: /mnt/workspace/yanhaiqiang.yhq/ChatLearn/output/grpo_test
+  data_path: /mnt/workspace/yanhaiqiang.yhq/ChatLearn/dataset/MATH-lighteval/train.json
+  eval_data_path: /mnt/workspace/yanhaiqiang.yhq/ChatLearn/dataset/MATH-lighteval/test.json
   data_checkpoint_path: ${runtime_args.output_dir}/data_checkpoint_path/
   # config for training
   num_episode: 200
-  sample_per_episode: 2048
+  sample_per_episode: 512
   train_micro_batch_size: 8
-  train_global_batch_size: 2048
+  train_global_batch_size: 512
   save_episode_interval: 200
   # config for data
   data_shuffle: True
@@ -21,7 +21,7 @@ runtime_args:
   eval_episode_interval: 5
   enable_eval_before_training: False
   log_args_dict:
-    log_dir: /mnt/workspace/yanhaiqiang.yhq/parameter_refactor/ChatLearn/output/grpo_test
+    log_dir: /mnt/workspace/yanhaiqiang.yhq/ChatLearn/output/grpo_test
 
 models:
   policy:
@@ -29,10 +29,10 @@ models:
       offload_weights: True
     generation_batch_size: 256
     gpu_per_process: 1
-    num_gpu: 8
+    num_gpu: 2
     trainable: False
     # args_dict:
-    tokenizer: /mnt/workspace/yanhaiqiang.yhq/parameter_refactor/ChatLearn/Qwen3-8B
+    tokenizer: /mnt/workspace/yanhaiqiang.yhq/hub/Qwen3-4B
     num_inference_per_prompt: 32
     seq_length: 2048
     max_seq_len_to_capture: 2348
@@ -55,8 +55,8 @@ models:
       offload_weights: True
     generation_batch_size: 8
     gpu_per_process: 1
-    num_gpu: 8
-    fsdp_size: 8
+    num_gpu: 2
+    fsdp_size: 2
     trainable: False
     # args_dict:
     # pretrain_or_model: /mnt/workspace/yanhaiqiang.yhq/parameter_refactor/ChatLearn/Qwen3-8B
@@ -67,10 +67,10 @@ models:
       offload_optimizer_states: True
       free_grad_buffers: True
     trainable: True
-    fsdp_size: 8
+    fsdp_size: 2
     generation_batch_size: 8
     gpu_per_process: 1
-    num_gpu: 8
+    num_gpu: 2
     # args_dict:
     # pretrain_or_model: /mnt/workspace/yanhaiqiang.yhq/parameter_refactor/ChatLearn/Qwen3-8B
     pretrain_or_model: ${models.policy.tokenizer}
